nutritionverse-tests/requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# API clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
aiohttp>=3.9.0

# Data storage
pyarrow>=14.0.0

# UI
streamlit>=1.30.0
plotly>=5.18.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter for FDC database

# Optional: for advanced features
# inflect>=7.0.0  # for better name normalization
